chore: improved error handling/reporting when document render step fails.
If an error occurs when attempting to render a document chunk, the
render function throws an excption (as opposed to just printing the
error). This was not being caught, which caused the whole application to
crash.

Now we catch the error and print out some information about the chunk
that failed to make it easier (hopefully) for the user to figure out
what went wrong...