break: replaced output_file_template argument with output_file
We now only support specifying the output file directly instead of a
template string that will be rendered with a context based on the input
file.

Added two new rules for automatic naming if the output_file argument is
not given. If the input file extension is .compudoc or .cd, then it is
removed.

example:
input          output
main.tex.cd -> main.tex